+++ examples/__init__.py
@@ -1,0 +1,1 @@
+"""Specialized, opt-in consumers of the generic standalone G-MES interface."""